More repo/service refactors
