+++ src/repositories/survey_choices_repository.py
@@ -47,5 +47,30 @@ class SurveyChoicesRepository:
         except Exception as e: # pylint: disable=W0718
             print(e)
             return False
+        
+    def create_new_choice_info(self, choice_id, info_key, info_value):
+        '''
+        Adds an additional to existing survey choice, updates choice_infos table
+        '''
+        try:
+            sql = "INSERT INTO choice_infos (choice_id, info_key, info_value)"\
+                " VALUES (:c_id, :i_key, :i_value)"
+            db.session.execute(text(sql), {"c_id":choice_id, "i_key":info_key, "i_value":info_value})
+            db.session.commit()
+        except Exception as e: # pylint: disable=W0718
+            print(e)
+            return False
+
+    def get_choice_additional_infos(self, choice_id):
+        '''
+        Gets a list of key-value pairs based on choice_id from choice_infos tables
+        '''
+        try:
+            sql = "SELECT info_key, info_value FROM choice_infos WHERE choice_id=:choice_id"
+            result = db.session.execute(text(sql), {"choice_id":choice_id})
+            return result.fetchall()
+        except Exception as e: # pylint: disable=W0718
+            print(e)
+            return False
 
 survey_choices_repository = SurveyChoicesRepository()
