+++ src/services/survey_choices_service.py
@@ -44,5 +44,8 @@ class SurveyChoicesService:
     
     def get_choice_name_and_spaces(self, choice_id):
         return self._choice_infos_repository.get_choice_name_and_spaces(choice_id)
+    
+    def get_choice_additional_infos(self, choice_id):
+            return self._choice_infos_repository.get_choice_additional_infos(choice_id)
 
 survey_choices_service = SurveyChoicesService()
